--- app/src/main/java/com/example/alexander/thevergetopnews/Components/IComponentProvider.java
@@ -1,8 +1,0 @@
-package com.example.alexander.thevergetopnews.Components;
-
-import com.example.alexander.thevergetopnews.Components.network.INetwork;
-
-public interface IComponentProvider {
-
-    INetwork getNetwork();
-}
