--- app/src/main/java/com/example/alexander/thevergetopnews/UI/Fragments/ListNewsFragment/AdapterCallback.java
@@ -1,8 +1,0 @@
-package com.example.alexander.thevergetopnews.UI.Fragments.ListNewsFragment;
-
-import com.example.alexander.thevergetopnews.Components.dto.Article;
-
-public interface AdapterCallback {
-
-    void onItemClick(Article article, int adapterPosition);
-}
