--- app/src/main/java/com/example/alexander/thevergetopnews/UI/Fragments/ListNewsFragment/TopicFragment/TopicFragmentContract.java
@@ -1,24 +1,0 @@
-package com.example.alexander.thevergetopnews.UI.Fragments.ListNewsFragment.TopicFragment;
-
-public class TopicFragmentContract {
-
-    public interface IPresenter {
-
-        void getData(String url);
-    }
-
-    public interface IView {
-
-        void showWebPage(String url);
-
-        void changeUI();
-    }
-     interface IFragment {
-
-    }
-
-    public interface IHost {
-
-        void hideElements(boolean isHidden);
-    }
-}
