+++ app/src/main/java/com/example/alexander/thevergetopnews/components/IComponentProvider.java
@@ -1,0 +1,8 @@
+package com.example.alexander.thevergetopnews.components;
+
+import com.example.alexander.thevergetopnews.components.network.INetwork;
+
+public interface IComponentProvider {
+
+    INetwork getNetwork();
+}
