+++ app/src/main/java/com/example/alexander/thevergetopnews/components/Injection.java
@@ -1,11 +1,11 @@
-package com.example.alexander.thevergetopnews.Components;
+package com.example.alexander.thevergetopnews.components;
 
-import com.example.alexander.thevergetopnews.UI.Fragments.ListNewsFragment.ListNewsFragmentPresenter;
-import com.example.alexander.thevergetopnews.UI.Fragments.ListNewsFragment.ListNewsFragmentContract;
-import com.example.alexander.thevergetopnews.UI.Activities.ScreenActivity.MainScreenPresenter;
-import com.example.alexander.thevergetopnews.UI.Activities.ScreenActivity.ScreenContract;
-import com.example.alexander.thevergetopnews.UI.Fragments.ListNewsFragment.TopicFragment.TopicFragmentContract;
-import com.example.alexander.thevergetopnews.UI.Fragments.ListNewsFragment.TopicFragment.TopicFragmentPresenter;
+import com.example.alexander.thevergetopnews.ui.news.MainScreenPresenter;
+import com.example.alexander.thevergetopnews.ui.news.ScreenContract;
+import com.example.alexander.thevergetopnews.ui.news.fr.article.TopicFragmentContract;
+import com.example.alexander.thevergetopnews.ui.news.fr.article.TopicFragmentPresenter;
+import com.example.alexander.thevergetopnews.ui.news.fr.news.ListNewsFragmentContract;
+import com.example.alexander.thevergetopnews.ui.news.fr.news.ListNewsFragmentPresenter;
 
 public class Injection {
 
@@ -24,7 +24,7 @@ public class Injection {
     }
 
     public static ListNewsFragmentContract.IPresenter getMainFragmentPresenter(ListNewsFragmentContract.IView view) {
-        return new ListNewsFragmentPresenter(view);
+        return new ListNewsFragmentPresenter(view, provider.getNetwork());
     }
 
     public static TopicFragmentContract.IPresenter  getTopicFragmentPresenter(TopicFragmentContract.IView view) {
