+++ app/src/main/java/com/example/alexander/thevergetopnews/components/network/NetworkImpl.java
@@ -1,6 +1,6 @@
-package com.example.alexander.thevergetopnews.Components.network;
+package com.example.alexander.thevergetopnews.components.network;
 
-import com.example.alexander.thevergetopnews.Components.dto.News;
+import com.example.alexander.thevergetopnews.components.dto.News;
 
 import java.util.concurrent.TimeUnit;
