+++ app/src/main/java/com/example/alexander/thevergetopnews/ui/news/fr/news/AdapterCallback.java
@@ -1,0 +1,8 @@
+package com.example.alexander.thevergetopnews.ui.news.fr.news;
+
+import com.example.alexander.thevergetopnews.components.dto.Article;
+
+public interface AdapterCallback {
+
+    void onItemClick(Article article, int adapterPosition);
+}
