+++ app/src/main/res/layout/activity_topic.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.Activities.TopicActivity.TopicActivity">
+    tools:context=".ui.Activities.TopicActivity.TopicActivity">
 
     <WebView
         android:id="@+id/webView"
